requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
docutils==0.18.1
Flask==2.1.0
Flask-SSLify==0.1.5
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
loguru==0.6.0
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.3
pydantic==1.9.0
Pygments==2.11.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
SQLAlchemy==1.4.31
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.1.0
yarl==1.7.2
zipp==3.7.0
